refactor: 패키지 이동 및 불필요한 테스트 제거
