+++ src/test/java/domain/board/BoardPositionTest.java
@@ -1,10 +1,8 @@
-package domain;
+package domain.board;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
 
-import domain.board.BoardPosition;
-import domain.board.Offset;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
